Create Facebook Boost Text Box Tool
If you at any time in your work have had to deal the the arbitrary voodoo that is the Facebook 20% text rule which determines whether or not a post can be boosted for cash money, the first thing I'd like to do is apologize. Neither the implementation of it, nor the reasons behind it, make a lick of sense and the whole process of having to hamstring your designs by strategically fitting text into boxes is painful.

The second thing I'd like to do is offer you this handy little Photoshop script that throws down guides that divide any image into said boxes; five vertically and five horizontally. After that the ball is in your court. I could write a book about the trials and tribulations of getting stuff boosted on Facebook. Good luck grasshopper.

p.s. this script should work in just about any version of Photoshop, just toss it in your scripts folder, reload PS and call it front the scripts menu.